+++ ft-sdk/src/from_request/app_url.rs
@@ -1,4 +1,4 @@
-/// Mountpoint is the url at which the app is installed using [`fastn.app`
+/// `ft_sdk::AppUrl` is the url at which the app is installed using [`fastn.app`
 /// feature](https://fastn.com/app/).
 ///
 /// If in FASTN.ftd, we have:
@@ -35,39 +35,66 @@
 /// url: /-/auth/
 /// ```
 ///
-/// If the wasm running for `my-app.com` needs to get the URL at which lets-auth app is installed
+/// If the wasm running for `my-app.com` needs to get the URL at which lets-auth app is installed,
 /// it will have to use `ft_sdk::AppUrl<"lets-auth">`.
 ///
-/// Implementation note: The `app_url` is passed by the host using `x-fastn-app-url` header.
-/// Host also passes `x-fastn-app-urls` containing app-urls of all the apps that are installed,
-/// and this app has access to.
-/// Some apps can be installed but may not be accessible to this app due to security reasons.
+/// Implementation note: The `app url` is passed by the host using `x-fastn-app-url` header. Host
+/// also passes `x-fastn-app-urls` containing app-urls of all the apps that are installed, and this
+/// app has access to. Some apps can be installed but may not be accessible to this app due to
+/// security reasons.
 pub struct AppUrl<const KEY: &'static str = CURRENT_APP_KEY>(pub Option<String>);
 
 pub const APP_URL_HEADER: &str = "x-fastn-app-url";
 pub const APP_URLS_HEADER: &str = "x-fastn-app-urls";
 pub const CURRENT_APP_KEY: &str = "current-app";
 
+impl<const KEY: &'static str> AppUrl<KEY> {
+    /// use this to combine app relative url with the app-url to construct full url
+    /// TODO: this should actually return full URI, including the query params etc
+    pub fn join(&self, path: &str) -> ft_sdk::Result<String> {
+        let v = match self.0 {
+            Some(ref v) => v,
+            None => return Err(anyhow::anyhow!("app-url not found for {KEY}")),
+        };
+
+        Ok(format!("{}{}/", v, path.trim_matches('/')))
+    }
+
+    pub fn is_set(&self) -> bool {
+        self.0.is_some()
+    }
+}
+
 impl<const KEY: &'static str> ft_sdk::FromRequest for AppUrl<KEY> {
-    fn from_request(req: &http::Request<serde_json::Value>) -> Result<AppUrl<KEY>, ft_sdk::Error> {
-        if KEY != CURRENT_APP_KEY {
-            Ok(Self(Some(
+    fn from_request(req: &http::Request<serde_json::Value>) -> ft_sdk::Result<AppUrl<KEY>> {
+        let v = if KEY != CURRENT_APP_KEY {
+            Some(
                 req.headers()
                     .get(APP_URL_HEADER)
                     .expect("host always provides this header")
                     .to_str()?
                     .to_string(),
-            )))
+            )
         } else {
-            Ok(Self(
-                serde_json::from_str::<std::collections::HashMap<String, String>>(
-                    req.headers()
-                        .get(APP_URLS_HEADER)
-                        .expect("host always provides this header")
-                        .to_str()?,
-                )?
-                .remove(KEY),
-            ))
+            serde_json::from_str::<std::collections::HashMap<String, String>>(
+                req.headers()
+                    .get(APP_URLS_HEADER)
+                    .expect("host always provides this header")
+                    .to_str()?,
+            )?
+            .remove(KEY)
+        };
+
+        if v.is_none() {
+            // this can happen either the application made an error and passed a wrong system name,
+            // or if the system is not installed, or if the system is installed, but this app does
+            // not have access to that system.
+            //
+            // to safeguard against the misspelling system name issue one should not hardcode the
+            // system name, but instead use the sdk of app they are trying to communicate with.
+            ft_sdk::println!("app-url not found for {KEY}");
         }
+
+        Ok(Self(v.map(|v| format!("/{}/", v.trim_matches('/')))))
     }
 }